interrupt io moved to external class, cop0 structs created

+++ src/device/device.h
@@ -2,7 +2,7 @@
 #include <stdint.h>
 
 namespace device {
-enum class Bit : uint8_t { cleared = 0, set = 1 };
+enum class Bit : uint32_t { cleared = 0, set = 1 };
 class Device {
    public:
     virtual ~Device(){};
